Additional fixes
Fixed

- `update` action now tries to update the toolchain image as well. Previously
this was not the case contrary to what someone would expect. Considering it won't
change the behaviour for non-latest images and the toolchain did not have any
breaking changes for a long time, this is not considered a breaking change either.
- Correct some errors' exit codes.

 Changed

- Deprecated providing the image flag for `install` action by displaying a
warning and removing it from documentation, without changing behaviour even
though it is higly unlikely this feature was ever used. It mainly exists for
historical reasons and it wil be removed in next major release.
- Update documentation to reflect `update` action changes.