fix(init.js): show a custom error when the submodule operation fails
This generally occurs to people because they copy the executable in their project folder. Even
though it has an extension on Windows, git thinks it is already being used as a prefix. I think this
might relate to how git is installed w.r.t the new experimental symlink support. In any case, we
speculate the user have the executable in the wrong place and show a relevant error.

fix #57