chore: improve problem matching and fix cpp server crash issue
I suspect because of the forcedInclude of the testrom.c on every compilation unit, the server was
crashing. This means that intellisense may not work on test files. OTOH, they seem to work ok in
general and this is something to improve on libdragon. The build times are high because it is a
single large c file already. Also fixed the c and cpp standards and the extension seem to correctly
parse almost everything.