fix(init.js): do not run git init unnecessarily
The tool depends on git for managing the vendored libdragon copy. It used to run `git init` to make
sure there is a git repos initialized. This works if the p;roject root matched the git troot. This
was mostly the case but when there is an NPM root we were attaching to it to be npm project
compatible and if that root didin't match the git root, this was cresulting in an extra git root.

re #69