feat(init.js): add a new --branch flag to override the libdragon branch
This is currently only implemented for the init action. It is possible that we also enable updating
it when running an `update`. The branch information is persisted on the configuration file mainly
because it is not recoverable from a `subtree` vendor. Still this change tries to recover it from a
submodule if the project gets destroyed. Also removes some unnecessary git ops from utils.

fixes #75