fix: run all commands in a shell and quote all user provided parameters
The other approach proved problematic as there were different commands still requiring a shell. Now
we always use the default shell but this has the drawback of not having quotes causing parsing
issues in the container. There are probably still issues with escape sequences in user provided
parameters, but that's for another day.